backend/requirements.txt
# Core Dependencies
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
starlette
slowapi

# File Processing
PyPDF2
pdfplumber

# HTTP Client
aiohttp
httpx

# LLM Dependencies
openai
requests

# Utilities
python-dateutil
typing-extensions
redis  # For rate limiting storage